requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
abseil-cpp=20211102.0=hd4dd3e8_0
absl-py=1.4.0=py39h06a4308_0
aiohttp=3.8.3=py39h5eee18b_0
aiosignal=1.2.0=pyhd3eb1b0_0
appdirs=1.4.4=pyhd3eb1b0_0
asttokens=2.0.5=pyhd3eb1b0_0
async-timeout=4.0.2=py39h06a4308_0
attrs=22.1.0=py39h06a4308_0
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
blinker=1.4=py39h06a4308_0
bottleneck=1.3.5=py39h7deecbd_0
brotli=1.0.9=h5eee18b_7
brotli-bin=1.0.9=h5eee18b_7
brotlipy=0.7.0=py39h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
c-ares=1.19.0=h5eee18b_0
ca-certificates=2023.05.30=h06a4308_0
cachetools=4.2.2=pyhd3eb1b0_0
certifi=2023.7.22=py39h06a4308_0
cffi=1.15.1=py39h5eee18b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.4=py39h06a4308_0
comm=0.1.2=py39h06a4308_0
contourpy=1.0.5=py39hdb19cb5_0
cryptography=41.0.2=py39h22a60cf_0
cuda-cudart=11.8.89=0
cuda-cupti=11.8.87=0
cuda-libraries=11.8.0=0
cuda-nvrtc=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-runtime=11.8.0=0
cycler=0.11.0=pyhd3eb1b0_0
debugpy=1.6.7=py39h6a678d5_0
decorator=5.1.1=pyhd3eb1b0_0
et-xmlfile=1.1.0=pypi_0
executing=0.8.3=pyhd3eb1b0_0
ffmpeg=4.3=hf484d3e_0
filelock=3.9.0=py39h06a4308_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.3.3=py39h5eee18b_0
fsspec=2023.4.0=py39h06a4308_0
giflib=5.2.1=h5eee18b_3
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py39heeb90bb_0
gnutls=3.6.15=he1e5248_0
google-auth=2.6.0=pyhd3eb1b0_0
google-auth-oauthlib=0.5.2=py39h06a4308_0
grpc-cpp=1.48.2=he1ff14a_1
grpcio=1.48.2=py39he1ff14a_1
idna=3.4=py39h06a4308_0
importlib-metadata=6.0.0=py39h06a4308_0
importlib_metadata=6.0.0=hd3eb1b0_0
importlib_resources=5.2.0=pyhd3eb1b0_1
intel-openmp=2023.1.0=hdb19cb5_46305
ipykernel=6.25.0=py39h2f386ee_0
ipython=8.12.0=py39h06a4308_0
jedi=0.18.1=py39h06a4308_1
jinja2=3.1.2=py39h06a4308_0
joblib=1.2.0=py39h06a4308_0
jpeg=9e=h5eee18b_1
jsonschema=4.17.3=py39h06a4308_0
jupyter_client=8.1.0=py39h06a4308_0
jupyter_core=5.3.0=py39h06a4308_0
kiwisolver=1.4.4=py39h6a678d5_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libbrotlicommon=1.0.9=h5eee18b_7
libbrotlidec=1.0.9=h5eee18b_7
libbrotlienc=1.0.9=h5eee18b_7
libcublas=11.11.3.6=0
libcufft=10.9.0.58=0
libcufile=1.7.1.12=0
libcurand=10.3.3.129=0
libcusolver=11.4.1.48=0
libcusparse=11.7.5.86=0
libdeflate=1.17=h5eee18b_0
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.4=h5eee18b_0
libnpp=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libpng=1.6.39=h5eee18b_0
libprotobuf=3.20.3=he621ea3_0
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.0=h6a678d5_2
libunistring=0.9.10=h27cfd23_0
libwebp=1.2.4=h11a3e52_1
libwebp-base=1.2.4=h5eee18b_1
lightning-utilities=0.7.1=py39h06a4308_1
lz4-c=1.9.4=h6a678d5_0
markdown=3.4.1=py39h06a4308_0
markupsafe=2.1.1=py39h7f8727e_0
matplotlib-base=3.7.1=py39h417a72b_1
matplotlib-inline=0.1.6=py39h06a4308_0
mkl=2023.1.0=h6d00ec8_46342
mkl-service=2.4.0=py39h5eee18b_1
mkl_fft=1.3.6=py39h417a72b_1
mkl_random=1.2.2=py39h417a72b_1
mlxtend=0.22.0=py39h06a4308_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.3.0=py39h06a4308_0
multidict=6.0.2=py39h5eee18b_0
munkres=1.1.4=py_0
nbformat=5.7.0=py39h06a4308_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.5.6=py39h06a4308_0
nettle=3.7.3=hbbd107a_1
networkx=3.1=py39h06a4308_0
numexpr=2.8.4=py39hc78ab66_1
numpy=1.25.0=py39h5f9d8c6_0
numpy-base=1.25.0=py39hb5e798b_0
oauthlib=3.2.2=py39h06a4308_0
openh264=2.1.1=h4ff587b_0
openpyxl=3.1.2=pypi_0
openssl=3.0.10=h7f8727e_0
packaging=23.0=py39h06a4308_0
pandas=1.5.3=py39h417a72b_0
parso=0.8.3=pyhd3eb1b0_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.4.0=py39h6a678d5_0
pip=23.2.1=py39h06a4308_0
platformdirs=2.5.2=py39h06a4308_0
plotly=5.9.0=py39h06a4308_0
pooch=1.4.0=pyhd3eb1b0_0
prompt-toolkit=3.0.36=py39h06a4308_0
protobuf=3.20.3=py39h6a678d5_0
psutil=5.9.0=py39h5eee18b_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
pyasn1=0.4.8=pyhd3eb1b0_0
pyasn1-modules=0.2.8=py_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.15.1=py39h06a4308_1
pyjwt=2.4.0=py39h06a4308_0
pyopenssl=23.2.0=py39h06a4308_0
pyparsing=3.0.9=py39h06a4308_0
pyrsistent=0.18.0=py39heee7806_0
pysocks=1.7.1=py39h06a4308_0
python=3.9.17=h955ad1f_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-fastjsonschema=2.16.2=py39h06a4308_0
pytorch=2.0.1=py3.9_cuda11.8_cudnn8.7.0_0
pytorch-cuda=11.8=h7e8668a_5
pytorch-lightning=2.0.3=py39h06a4308_0
pytorch-mutex=1.0=cuda
pytz=2022.7=py39h06a4308_0
pyyaml=6.0=py39h5eee18b_1
pyzmq=25.1.0=py39h6a678d5_0
re2=2022.04.01=h295c915_0
readline=8.2=h5eee18b_0
requests=2.31.0=py39h06a4308_0
requests-oauthlib=1.3.0=py_0
rsa=4.7.2=pyhd3eb1b0_1
scikit-learn=1.3.0=py39h1128e8f_0
scipy=1.10.1=py39hf6e8229_1
setuptools=68.0.0=py39h06a4308_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.41.2=h5eee18b_0
stack_data=0.2.0=pyhd3eb1b0_0
sympy=1.11.1=py39h06a4308_0
tbb=2021.8.0=hdb19cb5_0
tenacity=8.2.2=py39h06a4308_0
tensorboard=2.12.1=py39h06a4308_0
tensorboard-data-server=0.7.0=py39h52d8a92_0
tensorboard-plugin-wit=1.8.1=py39h06a4308_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.12=h1ccaba5_0
torchaudio=2.0.2=py39_cu118
torchmetrics=0.11.4=py39h2f386ee_1
torchtriton=2.0.0=py39
torchvision=0.15.2=py39_cu118
tornado=6.3.2=py39h5eee18b_0
tqdm=4.65.0=py39hb070fc8_0
traitlets=5.7.1=py39h06a4308_0
typing-extensions=4.7.1=py39h06a4308_0
typing_extensions=4.7.1=py39h06a4308_0
tzdata=2023c=h04d1e81_0
urllib3=1.26.16=py39h06a4308_0
wcwidth=0.2.5=pyhd3eb1b0_0
werkzeug=2.2.3=py39h06a4308_0
wheel=0.38.4=py39h06a4308_0
xz=5.4.2=h5eee18b_0
yaml=0.2.5=h7b6447c_0
yarl=1.8.1=py39h5eee18b_0
zeromq=4.3.4=h2531618_0
zipp=3.11.0=py39h06a4308_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
